V4 Create New Table
This revision was so exciting!!! I broke away from the tutorials and
tried to figure out how to create a new table specifically for the newly
registered user.  I did this so that all of the data I plan to store on
the database and associate with specific users is more organized. Now
justin1 will have all of the data for justin instead of looking through
users for find justin and all of his data.