V7 Day/Month/Year Data Display
I created stats.php to handle displaying the data in an organized
fashion.  This way I can just include the file where I want on several
pages.  The next and final feature I want to implement is a contacts
book.  So far I've been dealing with numbers, number are pretty easy to
work with.  I'm a little nervous about dealing with strings in the
database.  More learning to do though!